Rain 범위,크기 수정, Stat(Greed,Growth,Area)연동

+++ ProjectV/Assets/Scripts/Manager/DataManager.cs
@@ -5,7 +5,7 @@ public class CurrentGameData
 {
     public PlayerCharacterName characterName;
     public int killCount = 0;
-    public int gold = 0;
+    public float gold = 0;
     public float playTime = 0f;
 }
